Move mutating methods to ModelOps
