+++ core/api/src/models/Destination.ts
@@ -1,4 +1,4 @@
-import { task, api, log, config } from "actionhero";
+import { task, api } from "actionhero";
 import {
   Table,
   Column,
@@ -28,13 +28,13 @@ import { Export } from "./Export";
 import { Run } from "./Run";
 import { DestinationGroup } from "./DestinationGroup";
 import { DestinationGroupMembership } from "./DestinationGroupMembership";
-import { ExportProfilePluginMethod } from "../classes/plugin";
 import { plugin } from "../modules/plugin";
 import { Op } from "sequelize";
 import { OptionHelper } from "./../modules/optionHelper";
 import { MappingHelper } from "./../modules/mappingHelper";
 import { StateMachine } from "./../modules/stateMachine";
 import { ProfilePropertyRule } from "./ProfilePropertyRule";
+import { DestinationOps } from "./../modules/ops/destination";
 
 export interface DestinationMapping extends MappingHelper.Mappings {}
 export interface SimpleDestinationGroupMembership {
@@ -101,134 +101,6 @@ export class Destination extends LoggedModel<Destination> {
 
   @HasMany(() => Export)
   exports: Export[];
-
-  @BeforeSave
-  static async ensureUniqueName(instance: Destination) {
-    const count = await Destination.count({
-      where: {
-        guid: { [Op.ne]: instance.guid },
-        name: instance.name,
-        state: { [Op.ne]: "draft" },
-      },
-    });
-    if (count > 0) {
-      throw new Error(`name "${instance.name}" is already in use`);
-    }
-  }
-
-  @BeforeCreate
-  static async ensureAppReady(instance: Destination) {
-    const app = await App.findByGuid(instance.appGuid);
-    if (app.state !== "ready") {
-      throw new Error(`app ${app.guid} is not ready`);
-    }
-  }
-
-  @BeforeCreate
-  static async ensureExportProfilesMethod(instance: Destination) {
-    const { pluginConnection } = await instance.getPlugin();
-    if (!pluginConnection) {
-      throw new Error(`a destination of type ${instance.type} cannot be found`);
-    }
-    if (!pluginConnection.methods.exportProfile) {
-      throw new Error(
-        `a destination of type ${instance.type} cannot be created as there is no exportProfile method`
-      );
-    }
-  }
-
-  @BeforeSave
-  static async ensureOnlyOneDestinationPerApp(instance: Destination) {
-    const otherDestination = await Destination.scope(null).findOne({
-      where: {
-        appGuid: instance.appGuid,
-        guid: { [Op.not]: instance.guid },
-      },
-    });
-
-    if (otherDestination) {
-      throw new Error(
-        `destination "${otherDestination.name}" is already using this app`
-      );
-    }
-  }
-
-  @BeforeSave
-  static async updateState(instance: Destination) {
-    await StateMachine.transition(instance, STATE_TRANSITIONS);
-  }
-
-  @AfterSave
-  static async buildDestinationGroupsWhenTrackingAll(instance: Destination) {
-    if (instance.trackAllGroups === true) {
-      const groups = await Group.findAll();
-      for (const i in groups) {
-        await DestinationGroup.findOrCreate({
-          where: { groupGuid: groups[i].guid, destinationGuid: instance.guid },
-        });
-      }
-    }
-
-    if (!instance.changed()) {
-      return;
-    }
-
-    const changes = instance.changed() as Array<string>;
-    if (
-      instance.trackAllGroups === false &&
-      changes.includes("trackAllGroups")
-    ) {
-      const destinationGroups = await instance.$get("destinationGroups");
-      for (const i in destinationGroups) {
-        await destinationGroups[i].destroy();
-      }
-    }
-  }
-
-  @BeforeDestroy
-  static async cannotDeleteDestinationWithTrackkedGroups(
-    instance: Destination
-  ) {
-    const destinationGroupsCount = await instance.$count("destinationGroups");
-    if (instance.trackAllGroups === true || destinationGroupsCount > 0) {
-      throw new Error("cannot delete a destination that is tracking a group");
-    }
-  }
-
-  @AfterDestroy
-  static async destroyDestinationMappings(instance: Destination) {
-    return Mapping.destroy({
-      where: { ownerGuid: instance.guid },
-    });
-  }
-
-  @AfterDestroy
-  static async destroyDestinationOptions(instance: Destination) {
-    return Option.destroy({
-      where: { ownerGuid: instance.guid },
-    });
-  }
-
-  @AfterDestroy
-  static async destroyDestinationGroupMemberships(instance: Destination) {
-    return DestinationGroupMembership.destroy({
-      where: { destinationGuid: instance.guid },
-    });
-  }
-
-  @AfterDestroy
-  static async destroyDestinationGroups(instance: Destination) {
-    // need to go 1-by-1 for callbacks
-    const destinationGroups = await instance.$get("destinationGroups");
-    await Promise.all(destinationGroups.map((dsg) => dsg.destroy()));
-  }
-
-  @AfterDestroy
-  static async destroyExports(instance: Destination) {
-    await task.enqueue("destination:destroyExports", {
-      destinationGuid: instance.guid,
-    });
-  }
 
   async apiData(includeApp = true, includeGroups = true) {
     let app: App;
@@ -545,211 +417,21 @@ export class Destination extends LoggedModel<Destination> {
     newGroups: Array<Group>,
     sync = false
   ) {
-    const app = await this.$get("app");
-    let method: ExportProfilePluginMethod;
-    const { pluginConnection } = await this.getPlugin();
-    method = pluginConnection.methods.exportProfile;
-
-    if (!method) {
-      throw new Error(`cannot find an export method for app type ${app.type}`);
-    }
-
-    const appOptions = await app.getOptions();
-    await app.validateOptions(appOptions);
-    const destinationGroupMemberships = await this.getDestinationGroupMemberships();
-
-    const mapping = await this.getMapping();
-    const mappingKeys = Object.keys(mapping);
-    let mappedOldProfileProperties = {};
-    let mappedNewProfileProperties = {};
-    mappingKeys.forEach((k) => {
-      mappedOldProfileProperties[k] = oldProfileProperties[mapping[k]];
-      mappedNewProfileProperties[k] = newProfileProperties[mapping[k]];
-    });
-
-    const oldGroupNames = oldGroups
-      .filter((group) =>
-        destinationGroupMemberships
-          .map((dgm) => dgm.groupGuid)
-          .includes(group.guid)
-      )
-      .map(
-        (group) =>
-          destinationGroupMemberships.filter(
-            (dgm) => dgm.groupGuid === group.guid
-          )[0].remoteKey
-      );
-    const newGroupNames = newGroups
-      .filter((group) =>
-        destinationGroupMemberships
-          .map((dgm) => dgm.groupGuid)
-          .includes(group.guid)
-      )
-      .map(
-        (group) =>
-          destinationGroupMemberships.filter(
-            (dgm) => dgm.groupGuid === group.guid
-          )[0].remoteKey
-      );
-
-    const newGroupGuids = newGroups.map((g) => g.guid);
-    const destinationGroupGuids = (await this.$get("groups")).map(
-      (g) => g.guid
+    return DestinationOps.exportProfile(
+      this,
+      profile,
+      runs,
+      imports,
+      oldProfileProperties,
+      newProfileProperties,
+      oldGroups,
+      newGroups,
+      sync
     );
-    let toDelete = true;
-    newGroupGuids.forEach((newGroupGuid) => {
-      if (destinationGroupGuids.includes(newGroupGuid)) {
-        toDelete = false;
-      }
-    });
-
-    const mostRecentExport = await Export.findOne({
-      where: {
-        destinationGuid: this.guid,
-        profileGuid: profile.guid,
-        mostRecent: true,
-      },
-    });
-
-    if (mostRecentExport) {
-      const mostRecentMappedProfilePropertyKeys = Object.keys(
-        mostRecentExport.newProfileProperties
-      );
-      const currentMappedNewProfilePropertyKeys = Object.keys(
-        mappedNewProfileProperties
-      );
-      const currentMappedOldProfilePropertyKeys = Object.keys(
-        mappedNewProfileProperties
-      );
-
-      // since this export was previously mapped, we can assume the previous mapping still makes sense...
-      mostRecentMappedProfilePropertyKeys
-        .filter((k) => !currentMappedNewProfilePropertyKeys.includes(k))
-        .filter((k) => !currentMappedOldProfilePropertyKeys.includes(k))
-        .forEach((k) => (mappedOldProfileProperties[k] = "unknown"));
-    }
-
-    const _export = await Export.create({
-      destinationGuid: this.guid,
-      profileGuid: profile.guid,
-      runGuids: runs ? runs.map((run) => run.guid) : [],
-      startedAt: sync ? new Date() : undefined,
-      oldProfileProperties: mappedOldProfileProperties,
-      newProfileProperties: mappedNewProfileProperties,
-      oldGroups: oldGroupNames.sort(),
-      newGroups: newGroupNames.sort(),
-      toDelete,
-    });
-
-    if (runs)
-      await Promise.all(runs.map((run) => run.increment("exportsCreated")));
-
-    await _export.associateImports(imports);
-
-    if (sync) {
-      return this.sendExport(_export, sync);
-    } else {
-      await task.enqueue(
-        "export:send",
-        {
-          destinationGuid: this.guid,
-          exportGuid: _export.guid,
-        },
-        `exports:${app.type}`
-      );
-    }
   }
 
   async sendExport(_export: Export, sync = false) {
-    await _export.update({ startedAt: new Date() });
-
-    const options = await this.getOptions();
-    const app = await this.$get("app");
-    const appOptions = await app.getOptions();
-    const connection = await app.getConnection();
-    const profile = await _export.$get("profile");
-
-    let method: ExportProfilePluginMethod;
-    const { pluginConnection } = await this.getPlugin();
-    method = pluginConnection.methods.exportProfile;
-
-    const open = await app.checkAndUpdateParallelism("incr");
-    if (!open) {
-      const message = `parallelism limit reached for ${app.type}`;
-      if (sync) {
-        throw new Error(message);
-      } else {
-        log(message + ", re-enqueuing export ${_export.guid}");
-        return task.enqueueIn(
-          config.tasks.timeout + 1,
-          "export:send",
-          {
-            destinationGuid: this.guid,
-            exportGuid: _export.guid,
-          },
-          `exports:${app.type}`
-        );
-      }
-    }
-
-    try {
-      const { success, retryDelay, error } = await method({
-        connection,
-        app,
-        appOptions,
-        destination: this,
-        destinationOptions: options,
-        profile,
-        oldProfileProperties: _export.oldProfileProperties,
-        newProfileProperties: _export.newProfileProperties,
-        oldGroups: _export.oldGroups,
-        newGroups: _export.newGroups,
-        toDelete: _export.toDelete,
-      });
-
-      if (!success && retryDelay && !sync) {
-        return task.enqueueIn(
-          retryDelay,
-          "export:send",
-          {
-            destinationGuid: this.guid,
-            exportGuid: _export.guid,
-          },
-          `exports:${app.type}`
-        );
-      }
-
-      if (!success && retryDelay && sync) {
-        throw error;
-      }
-
-      await _export.update({ completedAt: new Date() });
-      await _export.markMostRecent();
-
-      if (_export.runGuids) {
-        for (const i in _export.runGuids) {
-          const run = await Run.findByGuid(_export.runGuids[i]);
-          await run.increment("profilesExported");
-        }
-      }
-
-      await app.checkAndUpdateParallelism("decr");
-      return { success, retryDelay, error };
-    } catch (error) {
-      _export.errorMessage = error.toString();
-      await _export.save();
-
-      if (_export.runGuids) {
-        for (const i in _export.runGuids) {
-          const run = await Run.findByGuid(_export.runGuids[i]);
-          await run.increment("profilesExported");
-        }
-      }
-
-      await app.checkAndUpdateParallelism("decr");
-      error.message = `error exporting profile ${profile.guid} to destination ${this.guid}: ${error}`;
-      throw error;
-    }
+    return DestinationOps.sendExport(this, _export, sync);
   }
 
   // --- Class Methods --- //
@@ -760,6 +442,134 @@ export class Destination extends LoggedModel<Destination> {
       throw new Error(`cannot find ${this.name} ${guid}`);
     }
     return instance;
+  }
+
+  @BeforeSave
+  static async ensureUniqueName(instance: Destination) {
+    const count = await Destination.count({
+      where: {
+        guid: { [Op.ne]: instance.guid },
+        name: instance.name,
+        state: { [Op.ne]: "draft" },
+      },
+    });
+    if (count > 0) {
+      throw new Error(`name "${instance.name}" is already in use`);
+    }
+  }
+
+  @BeforeCreate
+  static async ensureAppReady(instance: Destination) {
+    const app = await App.findByGuid(instance.appGuid);
+    if (app.state !== "ready") {
+      throw new Error(`app ${app.guid} is not ready`);
+    }
+  }
+
+  @BeforeCreate
+  static async ensureExportProfilesMethod(instance: Destination) {
+    const { pluginConnection } = await instance.getPlugin();
+    if (!pluginConnection) {
+      throw new Error(`a destination of type ${instance.type} cannot be found`);
+    }
+    if (!pluginConnection.methods.exportProfile) {
+      throw new Error(
+        `a destination of type ${instance.type} cannot be created as there is no exportProfile method`
+      );
+    }
+  }
+
+  @BeforeSave
+  static async ensureOnlyOneDestinationPerApp(instance: Destination) {
+    const otherDestination = await Destination.scope(null).findOne({
+      where: {
+        appGuid: instance.appGuid,
+        guid: { [Op.not]: instance.guid },
+      },
+    });
+
+    if (otherDestination) {
+      throw new Error(
+        `destination "${otherDestination.name}" is already using this app`
+      );
+    }
+  }
+
+  @BeforeSave
+  static async updateState(instance: Destination) {
+    await StateMachine.transition(instance, STATE_TRANSITIONS);
+  }
+
+  @AfterSave
+  static async buildDestinationGroupsWhenTrackingAll(instance: Destination) {
+    if (instance.trackAllGroups === true) {
+      const groups = await Group.findAll();
+      for (const i in groups) {
+        await DestinationGroup.findOrCreate({
+          where: { groupGuid: groups[i].guid, destinationGuid: instance.guid },
+        });
+      }
+    }
+
+    if (!instance.changed()) {
+      return;
+    }
+
+    const changes = instance.changed() as Array<string>;
+    if (
+      instance.trackAllGroups === false &&
+      changes.includes("trackAllGroups")
+    ) {
+      const destinationGroups = await instance.$get("destinationGroups");
+      for (const i in destinationGroups) {
+        await destinationGroups[i].destroy();
+      }
+    }
+  }
+
+  @BeforeDestroy
+  static async cannotDeleteDestinationWithTrackkedGroups(
+    instance: Destination
+  ) {
+    const destinationGroupsCount = await instance.$count("destinationGroups");
+    if (instance.trackAllGroups === true || destinationGroupsCount > 0) {
+      throw new Error("cannot delete a destination that is tracking a group");
+    }
+  }
+
+  @AfterDestroy
+  static async destroyDestinationMappings(instance: Destination) {
+    return Mapping.destroy({
+      where: { ownerGuid: instance.guid },
+    });
+  }
+
+  @AfterDestroy
+  static async destroyDestinationOptions(instance: Destination) {
+    return Option.destroy({
+      where: { ownerGuid: instance.guid },
+    });
+  }
+
+  @AfterDestroy
+  static async destroyDestinationGroupMemberships(instance: Destination) {
+    return DestinationGroupMembership.destroy({
+      where: { destinationGuid: instance.guid },
+    });
+  }
+
+  @AfterDestroy
+  static async destroyDestinationGroups(instance: Destination) {
+    // need to go 1-by-1 for callbacks
+    const destinationGroups = await instance.$get("destinationGroups");
+    await Promise.all(destinationGroups.map((dsg) => dsg.destroy()));
+  }
+
+  @AfterDestroy
+  static async destroyExports(instance: Destination) {
+    await task.enqueue("destination:destroyExports", {
+      destinationGuid: instance.guid,
+    });
   }
 
   /**
