+++ core/api/src/models/Event.ts
@@ -2,8 +2,6 @@ import {
   Table,
   Column,
   Index,
-  CreatedAt,
-  UpdatedAt,
   AllowNull,
   BeforeCreate,
   BelongsTo,
@@ -15,13 +13,10 @@ import { LoggedModel } from "../classes/loggedModel";
 import { Op } from "sequelize";
 import { EventData } from "./EventData";
 import { Profile } from "./Profile";
-import { ProfileProperty } from "./ProfileProperty";
-import {
-  ProfilePropertyRule,
-  ProfilePropertyRuleFiltersWithKey,
-} from "./ProfilePropertyRule";
+import { ProfilePropertyRuleFiltersWithKey } from "./ProfilePropertyRule";
 import * as uuid from "uuid";
 import { api, task } from "actionhero";
+import { EventOps } from "../modules/ops/event";
 
 @Table({ tableName: "events", paranoid: false })
 export class Event extends LoggedModel<Event> {
@@ -70,27 +65,6 @@ export class Event extends LoggedModel<Event> {
   @BelongsTo(() => Profile, "profileGuid")
   profile: Profile;
 
-  @BeforeCreate
-  static generateGuid(instance: Event) {
-    if (!instance.guid) {
-      instance.guid = `${instance.guidPrefix()}_${uuid.v4()}`;
-    }
-  }
-
-  @AfterCreate
-  static async enqueueAssociateEvent(instance: Event) {
-    await task.enqueue("event:associateProfile", { eventGuid: instance.guid });
-  }
-
-  @AfterDestroy
-  static async destroyOptions(instance: Event) {
-    return EventData.destroy({
-      where: {
-        eventGuid: instance.guid,
-      },
-    });
-  }
-
   async getData() {
     const eventData = this.eventData || (await this.$get("eventData"));
     const data = {};
@@ -116,115 +90,7 @@ export class Event extends LoggedModel<Event> {
   }
 
   async associate(identifyingProfilePropertyRuleGuid: string) {
-    // find the cached profile property rule
-    const profilePropertyRules = await ProfilePropertyRule.cached();
-    const profilePropertyRuleCacheKey = Object.keys(
-      profilePropertyRules
-    ).filter(
-      (key) =>
-        profilePropertyRules[key].guid === identifyingProfilePropertyRuleGuid
-    )[0];
-    const profilePropertyRule =
-      profilePropertyRules[profilePropertyRuleCacheKey];
-
-    if (this.profileGuid) {
-      // we are already identified
-      try {
-        const profile = await this.$get("profile");
-        await this.updateProfile(profile);
-        return profile;
-      } catch (error) {
-        // the event may have been moved to another profile
-        const updatedEvent = await this.reload();
-        if (updatedEvent.profileGuid !== this.profileGuid) {
-          return updatedEvent.associate(updatedEvent.profileGuid);
-        } else {
-          throw error;
-        }
-      }
-    } else if (this.userId) {
-      // we have a userId (primaryIdentifyingProfilePropertyRule)
-      let profile = await Profile.findOne({
-        include: [
-          {
-            model: ProfileProperty,
-            where: {
-              rawValue: this.userId,
-              profilePropertyRuleGuid: profilePropertyRule.guid,
-            },
-          },
-        ],
-      });
-
-      if (!profile) {
-        [profile] = await Profile.findOrCreate({
-          where: { anonymousId: this.anonymousId },
-        });
-      }
-
-      const profileProperties = {};
-      profileProperties[profilePropertyRule.key] = this.userId;
-      await profile.addOrUpdateProperties(profileProperties);
-
-      this.profileGuid = profile.guid;
-      this.profileAssociatedAt = new Date();
-      await this.save();
-
-      const otherProfileWithAnonymousId = await Profile.findOne({
-        where: {
-          guid: { [Op.ne]: profile.guid },
-          anonymousId: this.anonymousId,
-        },
-      });
-      if (!otherProfileWithAnonymousId) {
-        await profile.update({ anonymousId: this.anonymousId });
-      } else {
-        await Profile.merge(profile, otherProfileWithAnonymousId);
-      }
-
-      await this.updateProfile(profile);
-      return profile;
-    } else if (this.anonymousId) {
-      // we have an anonymousId
-
-      // can we find the profile by anonymousId?
-      let profile = await Profile.findOne({
-        where: { anonymousId: this.anonymousId },
-      });
-
-      // can we find the profile from other events with the same anonymousId?
-      if (!profile) {
-        const event = await Event.findOne({
-          where: {
-            anonymousId: this.anonymousId,
-            profileGuid: { [Op.ne]: null },
-          },
-        });
-        if (event) {
-          profile = await Profile.findOne({
-            where: { guid: event.profileGuid },
-          });
-        }
-      }
-
-      // if we still don't have a profile, make a new one
-      if (!profile) {
-        [profile] = await Profile.findOrCreate({
-          where: { anonymousId: this.anonymousId },
-        });
-      }
-
-      this.profileGuid = profile.guid;
-      this.profileAssociatedAt = new Date();
-      await this.save();
-      await this.updateProfile(profile);
-      return profile;
-    } else {
-      // we can't identify this event
-      throw new Error(
-        "cannot associate a profile without profileGuid, userId, or anonymousId"
-      );
-    }
+    return EventOps.associate(this, identifyingProfilePropertyRuleGuid);
   }
 
   async updateProfile(profile: Profile) {
@@ -253,6 +119,27 @@ export class Event extends LoggedModel<Event> {
   }
 
   // --- Class Methods --- //
+
+  @BeforeCreate
+  static generateGuid(instance: Event) {
+    if (!instance.guid) {
+      instance.guid = `${instance.guidPrefix()}_${uuid.v4()}`;
+    }
+  }
+
+  @AfterCreate
+  static async enqueueAssociateEvent(instance: Event) {
+    await task.enqueue("event:associateProfile", { eventGuid: instance.guid });
+  }
+
+  @AfterDestroy
+  static async destroyOptions(instance: Event) {
+    return EventData.destroy({
+      where: {
+        eventGuid: instance.guid,
+      },
+    });
+  }
 
   static async findByGuid(guid: string) {
     const instance = await this.scope(null).findOne({ where: { guid } });
