+++ core/api/src/models/Group.ts
@@ -15,7 +15,7 @@ import {
   DefaultScope,
   Scopes,
 } from "sequelize-typescript";
-import { api, task } from "actionhero";
+import { api } from "actionhero";
 import { Op } from "sequelize";
 import Moment from "moment";
 import { LoggedModel } from "../classes/loggedModel";
@@ -23,9 +23,7 @@ import { GroupMember } from "./GroupMember";
 import { GroupRule } from "./GroupRule";
 import { Run } from "./Run";
 import { Profile } from "./Profile";
-import { Import } from "./Import";
 import { Setting } from "./Setting";
-import { ProfileMultipleAssociationShim } from "./ProfileMultipleAssociationShim";
 import { ProfileProperty } from "./ProfileProperty";
 import { Destination } from "./Destination";
 import { DestinationGroup } from "./DestinationGroup";
@@ -33,9 +31,10 @@ import { DestinationGroupMembership } from "./DestinationGroupMembership";
 import {
   ProfilePropertyRule,
   profilePropertyRuleJSToSQLType,
-  ProfilePropertyRuleOps,
 } from "./ProfilePropertyRule";
+import { ProfilePropertyRuleOpsDictionary } from "../modules/RuleOpsDictionary";
 import { StateMachine } from "./../modules/stateMachine";
+import { GroupOps } from "../modules/ops/group";
 
 export const GROUP_RULE_LIMIT = 10;
 const numbers = [...Array(GROUP_RULE_LIMIT).keys()].map((n) => n + 1).reverse();
@@ -134,112 +133,6 @@ export class Group extends LoggedModel<Group> {
   @BelongsToMany(() => Profile, () => GroupMember)
   profiles: Profile[];
 
-  @BeforeSave
-  static async ensureUniqueName(instance: Group) {
-    const count = await Group.count({
-      where: {
-        guid: { [Op.ne]: instance.guid },
-        name: instance.name,
-        state: { [Op.ne]: "draft" },
-      },
-    });
-    if (count > 0) {
-      throw new Error(`name "${instance.name}" is already in use`);
-    }
-  }
-
-  @BeforeSave
-  static async updateState(instance: Group) {
-    await StateMachine.transition(instance, STATE_TRANSITIONS);
-  }
-
-  @AfterSave
-  static async linkToDestinationsTrackingAllGroups(instance: Group) {
-    if (["draft", "deleted"].includes(instance.state)) return;
-
-    const destinations = await Destination.findAll({
-      where: { trackAllGroups: true },
-    });
-
-    for (const i in destinations) {
-      await DestinationGroup.findOrCreate({
-        where: {
-          groupGuid: instance.guid,
-          destinationGuid: destinations[i].guid,
-        },
-      });
-    }
-  }
-
-  @BeforeDestroy
-  static async checkGroupMembers(instance: Group) {
-    const count = await instance.$count("groupMembers");
-    if (count > 0) {
-      throw new Error(`this group still has ${count} members, cannot delete`);
-    }
-  }
-
-  @BeforeDestroy
-  static async checkDestinationGroups(instance: Group) {
-    const count = await DestinationGroup.count({
-      where: { groupGuid: instance.guid },
-      include: [{ model: Destination, where: { trackAllGroups: false } }],
-    });
-
-    if (count > 0) {
-      throw new Error(
-        `this group still in use by ${count} destinations, cannot delete`
-      );
-    }
-  }
-
-  @BeforeDestroy
-  static async checkDestinationGroupMembership(instance: Group) {
-    const count = await DestinationGroupMembership.count({
-      where: { groupGuid: instance.guid },
-    });
-
-    if (count > 0) {
-      throw new Error(
-        `this group still in use by ${count} destinations, cannot delete`
-      );
-    }
-  }
-
-  @AfterDestroy
-  static async destroyDestinationGroup(instance: Group) {
-    // need to go 1-by-1 for callbacks
-    const destinationGroups = await instance.$get("destinationGroups");
-    for (const i in destinationGroups) {
-      await destinationGroups[i].destroy();
-    }
-  }
-
-  @AfterDestroy
-  static async destroyDestinationGroupMembership(instance: Group) {
-    return DestinationGroupMembership.destroy({
-      where: {
-        groupGuid: instance.guid,
-      },
-    });
-  }
-
-  @AfterDestroy
-  static async destroyGroupRules(instance: Group) {
-    return GroupRule.destroy({
-      where: {
-        groupGuid: instance.guid,
-      },
-    });
-  }
-
-  @AfterDestroy
-  static async destroyRuns(instance: Group) {
-    await Run.destroy({
-      where: { creatorGuid: instance.guid },
-    });
-  }
-
   async profilesCount(options = {}) {
     let search = { where: { groupGuid: this.guid } };
     if (options) {
@@ -267,9 +160,9 @@ export class Group extends LoggedModel<Group> {
         type: profilePropertyRule.type,
         operation: {
           op: rule.op,
-          description: ProfilePropertyRuleOps[profilePropertyRule.type].filter(
-            (operation) => operation.op === rule.op
-          )[0].description,
+          description: ProfilePropertyRuleOpsDictionary[
+            profilePropertyRule.type
+          ].filter((operation) => operation.op === rule.op)[0].description,
         },
         match: rule.match,
         relativeMatchNumber: rule.relativeMatchNumber,
@@ -362,7 +255,7 @@ export class Group extends LoggedModel<Group> {
   }
 
   fromConvenientRules(rules: GroupRuleWithKey[]) {
-    const convenientRules = ProfilePropertyRuleOps._convenientRules;
+    const convenientRules = ProfilePropertyRuleOpsDictionary._convenientRules;
 
     for (const i in rules) {
       if (convenientRules[rules[i].operation.op]) {
@@ -385,7 +278,7 @@ export class Group extends LoggedModel<Group> {
   }
 
   toConvenientRules(rules: GroupRuleWithKey[]) {
-    const convenientRules = ProfilePropertyRuleOps._convenientRules;
+    const convenientRules = ProfilePropertyRuleOpsDictionary._convenientRules;
 
     for (const i in rules) {
       for (const k in convenientRules) {
@@ -403,7 +296,7 @@ export class Group extends LoggedModel<Group> {
         !rules[i].operation.op.match(/^relative_/)
       ) {
         rules[i].operation.op = `relative_${rules[i].operation.op}`;
-        rules[i].operation.description = ProfilePropertyRuleOps[
+        rules[i].operation.description = ProfilePropertyRuleOpsDictionary[
           rules[i].type
         ].filter(
           (operation) => operation.op === rules[i].operation.op
@@ -412,28 +305,6 @@ export class Group extends LoggedModel<Group> {
     }
 
     return rules;
-  }
-
-  async run(force = false, destinationGuid?: string) {
-    await this.stopPreviousRuns();
-    await task.enqueue("group:run", {
-      groupGuid: this.guid,
-      force,
-      destinationGuid,
-    });
-  }
-
-  async stopPreviousRuns() {
-    const previousRuns = await Run.findAll({
-      where: {
-        creatorGuid: this.guid,
-        state: "running",
-      },
-    });
-
-    for (const i in previousRuns) {
-      await previousRuns[i].stop();
-    }
   }
 
   async nextCalculatedAt() {
@@ -448,36 +319,30 @@ export class Group extends LoggedModel<Group> {
     return Moment(this.calculatedAt).add(delayMinutes, "minutes").toDate();
   }
 
-  async _buildProfileImport(
+  async run(force = false, destinationGuid?: string) {
+    return GroupOps.run(this, force, destinationGuid);
+  }
+
+  async stopPreviousRuns() {
+    return GroupOps.stopPreviousRuns(this);
+  }
+
+  async buildProfileImport(
     profileGuid: string,
     creatorType: string,
     creatorGuid: string,
     destinationGuid?: string
   ) {
-    const profile = await Profile.findOne({ where: { guid: profileGuid } });
-
-    const oldProfileProperties = [];
-    const properties = await profile.properties();
-    for (const key in properties) {
-      oldProfileProperties[key] = properties[key].value;
-    }
-
-    const oldGroupGuids = (await profile.$get("groups")).map((g) => g.guid);
-
-    return Import.build({
-      rawData: destinationGuid ? { _meta: { destinationGuid } } : {},
-      data: destinationGuid ? { _meta: { destinationGuid } } : {},
+    return GroupOps.buildProfileImport(
+      profileGuid,
       creatorType,
       creatorGuid,
-      profileGuid: profile.guid,
-      profileAssociatedAt: new Date(),
-      oldProfileProperties,
-      oldGroupGuids,
-    });
+      destinationGuid
+    );
   }
 
   async addProfile(profile: Profile) {
-    const _import = await this._buildProfileImport(
+    const _import = await this.buildProfileImport(
       profile.guid,
       "group",
       this.guid
@@ -500,7 +365,7 @@ export class Group extends LoggedModel<Group> {
       throw new Error("profile is not a member of this group");
     }
 
-    const _import = await this._buildProfileImport(
+    const _import = await this.buildProfileImport(
       profile.guid,
       "group",
       this.guid
@@ -518,74 +383,14 @@ export class Group extends LoggedModel<Group> {
     force = false,
     destinationGuid?: string
   ) {
-    let groupMembersCount = 0;
-    let profiles: ProfileMultipleAssociationShim[];
-
-    if (this.type === "manual") {
-      profiles = await this.$get("profiles", {
-        attributes: ["guid"],
-        limit,
-        offset,
-        order: [["createdAt", "asc"]],
-      });
-    } else {
-      const rules = await this.getRules();
-      if (Object.keys(rules).length === 0) {
-        return 0;
-      }
-
-      const { where, include } = await this._buildGroupMemberQueryParts(
-        rules,
-        this.matchType
-      );
-
-      profiles = await ProfileMultipleAssociationShim.findAll({
-        attributes: ["guid"],
-        where,
-        include,
-        limit,
-        offset,
-        order: [["createdAt", "asc"]],
-        subQuery: false,
-      });
-    }
-
-    for (const i in profiles) {
-      const profile = profiles[i];
-      const groupMember = await GroupMember.findOne({
-        where: {
-          profileGuid: profile.guid,
-          groupGuid: this.guid,
-        },
-      });
-
-      if (!groupMember || force) {
-        const transaction = await api.sequelize.transaction();
-        const _import = await this._buildProfileImport(
-          profile.guid,
-          "run",
-          run.guid,
-          destinationGuid
-        );
-        await run.increment(["importsCreated"], { transaction });
-        await _import.save({ transaction });
-        await transaction.commit();
-      }
-
-      if (groupMember) {
-        groupMember.removedAt = null;
-        groupMember.set("updatedAt", new Date());
-        groupMember.changed("updatedAt", true);
-        await groupMember.save();
-      }
-
-      groupMembersCount++;
-    }
-
-    this.calculatedAt = new Date();
-    await this.save();
-
-    return groupMembersCount;
+    return GroupOps.runAddGroupMembers(
+      this,
+      run,
+      limit,
+      offset,
+      force,
+      destinationGuid
+    );
   }
 
   async runRemoveGroupMembers(
@@ -595,177 +400,36 @@ export class Group extends LoggedModel<Group> {
     force = false,
     destinationGuid?: string
   ) {
-    let groupMembersCount = 0;
-
-    // The offset and order can be ignored as we will keep running this query until the set of results becomes 0.
-    const groupMembersToRemove = await GroupMember.findAll({
-      attributes: ["guid", "profileGuid"],
-      where: {
-        groupGuid: this.guid,
-        updatedAt: { [Op.lt]: run.createdAt },
-        removedAt: null,
-      },
+    return GroupOps.runRemoveGroupMembers(
+      this,
+      run,
       limit,
-    });
-
-    for (const i in groupMembersToRemove) {
-      const member = groupMembersToRemove[i];
-      const transaction = await api.sequelize.transaction();
-      member.removedAt = new Date();
-      await member.save({ transaction });
-      const _import = await this._buildProfileImport(
-        member.profileGuid,
-        "run",
-        run.guid,
-        destinationGuid
-      );
-      await run.increment(["importsCreated"], { transaction });
-      await _import.save({ transaction });
-      await transaction.commit();
-      groupMembersCount++;
-    }
-
-    this.calculatedAt = new Date();
-    await this.save();
-
-    return groupMembersCount;
+      offset,
+      force,
+      destinationGuid
+    );
   }
 
   async removePreviousRunGroupMembers(run: Run, limit = 100) {
-    let groupMembersCount = 0;
-
-    const groupMembersToRemove = await GroupMember.findAll({
-      attributes: ["guid", "profileGuid"],
-      where: {
-        groupGuid: this.guid,
-        removedAt: { [Op.lte]: run.createdAt },
-      },
-      limit,
-    });
-
-    for (const i in groupMembersToRemove) {
-      const member = groupMembersToRemove[i];
-      member.removedAt = new Date();
-      await member.save();
-      const _import = await this._buildProfileImport(
-        member.profileGuid,
-        "run",
-        run.guid
-      );
-      await _import.save();
-      groupMembersCount++;
-    }
-
-    return groupMembersCount;
+    return GroupOps.removePreviousRunGroupMembers(this, run, limit);
   }
 
   async updateProfileMembership(profile: Profile) {
-    const existingMembership = await GroupMember.findOne({
-      where: {
-        groupGuid: this.guid,
-        profileGuid: profile.guid,
-      },
-    });
-
-    if (this.type === "manual") {
-      if (this.state === "deleted") {
-        if (existingMembership) {
-          await existingMembership.destroy();
-        }
-        return false;
-      } else {
-        return existingMembership ? true : false;
-      }
-    } else {
-      const rules = await this.getRules();
-
-      if (Object.keys(rules).length == 0) {
-        if (existingMembership) {
-          await existingMembership.destroy();
-        }
-        return false;
-      } else {
-        const { where, include } = await this._buildGroupMemberQueryParts(
-          rules,
-          this.matchType
-        );
-
-        // and includes this profile
-        if (!where[Op.and]) {
-          where[Op.and] = [];
-        }
-        where[Op.and].push({
-          guid: profile.guid,
-        });
-
-        const matchedProfiles = await ProfileMultipleAssociationShim.findAll({
-          attributes: ["guid"],
-          where,
-          include,
-        });
-
-        const belongs = matchedProfiles.length === 1;
-
-        if (belongs && !existingMembership) {
-          await GroupMember.create({
-            groupGuid: this.guid,
-            profileGuid: profile.guid,
-          });
-        }
-
-        if (!belongs && existingMembership) {
-          await existingMembership.destroy();
-        }
-
-        return belongs;
-      }
-    }
+    return GroupOps.updateProfileMembership(this, profile);
   }
 
   async countPotentialMembers(
     rules?: GroupRuleWithKey[],
     matchType: "any" | "all" = this.matchType
   ) {
-    if (!rules) {
-      rules = await this.getRules();
-    }
-
-    const { where, include } = await this._buildGroupMemberQueryParts(
-      rules,
-      matchType
-    );
-
-    // we use ProfileMultipleAssociationShim as a shim model which has extra associations
-    return ProfileMultipleAssociationShim.count({ where, include });
+    return GroupOps.countPotentialMembers(this, rules, matchType);
   }
 
   /**
    * Take the rules for the group and check how many group members there would have been both individually for each single rule, and then by building up the query step-by-step
    */
   async countComponentMembersFromRules(rules?: GroupRuleWithKey[]) {
-    const componentCounts = [];
-    const funnelCounts = [];
-    let funnelStep = 0;
-
-    if (!rules) {
-      rules = await this.getRules();
-    }
-
-    for (const i in rules) {
-      const localRule = rules[i];
-      componentCounts[i] = await this.countPotentialMembers([localRule]);
-
-      const funnelRules = [];
-      let j = 0;
-      while (j <= funnelStep) {
-        funnelRules.push(rules[j]);
-        j++;
-      }
-      funnelCounts[funnelStep] = await this.countPotentialMembers(funnelRules);
-      funnelStep++;
-    }
-
-    return { componentCounts, funnelCounts };
+    return GroupOps.countComponentMembersFromRules(this, rules);
   }
 
   async _buildGroupMemberQueryParts(
@@ -883,5 +547,111 @@ export class Group extends LoggedModel<Group> {
       throw new Error(`cannot find ${this.name} ${guid}`);
     }
     return instance;
+  }
+
+  @BeforeSave
+  static async ensureUniqueName(instance: Group) {
+    const count = await Group.count({
+      where: {
+        guid: { [Op.ne]: instance.guid },
+        name: instance.name,
+        state: { [Op.ne]: "draft" },
+      },
+    });
+    if (count > 0) {
+      throw new Error(`name "${instance.name}" is already in use`);
+    }
+  }
+
+  @BeforeSave
+  static async updateState(instance: Group) {
+    await StateMachine.transition(instance, STATE_TRANSITIONS);
+  }
+
+  @AfterSave
+  static async linkToDestinationsTrackingAllGroups(instance: Group) {
+    if (["draft", "deleted"].includes(instance.state)) return;
+
+    const destinations = await Destination.findAll({
+      where: { trackAllGroups: true },
+    });
+
+    for (const i in destinations) {
+      await DestinationGroup.findOrCreate({
+        where: {
+          groupGuid: instance.guid,
+          destinationGuid: destinations[i].guid,
+        },
+      });
+    }
+  }
+
+  @BeforeDestroy
+  static async checkGroupMembers(instance: Group) {
+    const count = await instance.$count("groupMembers");
+    if (count > 0) {
+      throw new Error(`this group still has ${count} members, cannot delete`);
+    }
+  }
+
+  @BeforeDestroy
+  static async checkDestinationGroups(instance: Group) {
+    const count = await DestinationGroup.count({
+      where: { groupGuid: instance.guid },
+      include: [{ model: Destination, where: { trackAllGroups: false } }],
+    });
+
+    if (count > 0) {
+      throw new Error(
+        `this group still in use by ${count} destinations, cannot delete`
+      );
+    }
+  }
+
+  @BeforeDestroy
+  static async checkDestinationGroupMembership(instance: Group) {
+    const count = await DestinationGroupMembership.count({
+      where: { groupGuid: instance.guid },
+    });
+
+    if (count > 0) {
+      throw new Error(
+        `this group still in use by ${count} destinations, cannot delete`
+      );
+    }
+  }
+
+  @AfterDestroy
+  static async destroyDestinationGroup(instance: Group) {
+    // need to go 1-by-1 for callbacks
+    const destinationGroups = await instance.$get("destinationGroups");
+    for (const i in destinationGroups) {
+      await destinationGroups[i].destroy();
+    }
+  }
+
+  @AfterDestroy
+  static async destroyDestinationGroupMembership(instance: Group) {
+    return DestinationGroupMembership.destroy({
+      where: {
+        groupGuid: instance.guid,
+      },
+    });
+  }
+
+  @AfterDestroy
+  static async destroyGroupRules(instance: Group) {
+    return GroupRule.destroy({
+      where: {
+        groupGuid: instance.guid,
+      },
+    });
+  }
+
+  @AfterDestroy
+  static async destroyRuns(instance: Group) {
+    await Run.destroy({
+      where: { creatorGuid: instance.guid },
+    });
   }
 }
