+++ core/api/src/models/Team.ts
@@ -4,7 +4,6 @@ import {
   AllowNull,
   Default,
   Length,
-  AfterCreate,
   AfterSave,
   BeforeDestroy,
   HasMany,
@@ -45,6 +44,68 @@ export class Team extends LoggedModel<Team> {
 
   @HasMany(() => Permission)
   permissions: Permission[];
+
+  async apiData() {
+    const membersCount = await this.$count("teamMembers");
+    const permissions = await this.$get("permissions", {
+      order: [["topic", "asc"]],
+    });
+    const permissionsApiData: AsyncReturnType<
+      Permission["apiData"]
+    >[] = await Promise.all(permissions.map((prm) => prm.apiData()));
+
+    return {
+      guid: this.guid,
+      name: this.name,
+      locked: this.locked,
+      permissionAllRead: this.permissionAllRead,
+      permissionAllWrite: this.permissionAllWrite,
+      createdAt: this.createdAt ? this.createdAt.getTime() : null,
+      updatedAt: this.updatedAt ? this.updatedAt.getTime() : null,
+      permissions: permissionsApiData,
+      membersCount,
+    };
+  }
+
+  async authorizeAction(topic: string, mode: "read" | "write") {
+    return Permission.authorizeAction(this.guid, topic, mode);
+  }
+
+  async setPermissions(
+    permissions: Array<{ guid: string; read: boolean; write: boolean }>
+  ) {
+    for (const i in permissions) {
+      const permission = await Permission.findOne({
+        where: { ownerGuid: this.guid, guid: permissions[i].guid },
+      });
+      if (!permission) {
+        throw new Error(
+          `permission ${permissions[i].guid} not found for this team`
+        );
+      }
+      if (!permission.locked) {
+        permission.read =
+          this.permissionAllRead !== null
+            ? this.permissionAllRead
+            : permissions[i].read;
+        permission.write =
+          this.permissionAllWrite !== null
+            ? this.permissionAllWrite
+            : permissions[i].write;
+        await permission.save();
+      }
+    }
+  }
+
+  // --- Class Methods --- //
+
+  static async findByGuid(guid: string) {
+    const instance = await this.scope(null).findOne({ where: { guid } });
+    if (!instance) {
+      throw new Error(`cannot find ${this.name} ${guid}`);
+    }
+    return instance;
+  }
 
   @BeforeSave
   static async checkLockedPermissions(instance: Team) {
@@ -112,67 +173,5 @@ export class Team extends LoggedModel<Team> {
   @AfterDestroy
   static async deletePermissions(instance: Team) {
     return Permission.destroy({ where: { ownerGuid: instance.guid } });
-  }
-
-  async apiData() {
-    const membersCount = await this.$count("teamMembers");
-    const permissions = await this.$get("permissions", {
-      order: [["topic", "asc"]],
-    });
-    const permissionsApiData: AsyncReturnType<
-      Permission["apiData"]
-    >[] = await Promise.all(permissions.map((prm) => prm.apiData()));
-
-    return {
-      guid: this.guid,
-      name: this.name,
-      locked: this.locked,
-      permissionAllRead: this.permissionAllRead,
-      permissionAllWrite: this.permissionAllWrite,
-      createdAt: this.createdAt ? this.createdAt.getTime() : null,
-      updatedAt: this.updatedAt ? this.updatedAt.getTime() : null,
-      permissions: permissionsApiData,
-      membersCount,
-    };
-  }
-
-  async authorizeAction(topic: string, mode: "read" | "write") {
-    return Permission.authorizeAction(this.guid, topic, mode);
-  }
-
-  async setPermissions(
-    permissions: Array<{ guid: string; read: boolean; write: boolean }>
-  ) {
-    for (const i in permissions) {
-      const permission = await Permission.findOne({
-        where: { ownerGuid: this.guid, guid: permissions[i].guid },
-      });
-      if (!permission) {
-        throw new Error(
-          `permission ${permissions[i].guid} not found for this team`
-        );
-      }
-      if (!permission.locked) {
-        permission.read =
-          this.permissionAllRead !== null
-            ? this.permissionAllRead
-            : permissions[i].read;
-        permission.write =
-          this.permissionAllWrite !== null
-            ? this.permissionAllWrite
-            : permissions[i].write;
-        await permission.save();
-      }
-    }
-  }
-
-  // --- Class Methods --- //
-
-  static async findByGuid(guid: string) {
-    const instance = await this.scope(null).findOne({ where: { guid } });
-    if (!instance) {
-      throw new Error(`cannot find ${this.name} ${guid}`);
-    }
-    return instance;
   }
 }
